apply functional programming
Instead of calculating all values in advance, prepare the new value directly for the plotter. This eliminates the need for an intermediate buffer before writing to the temporary file.

To allow different simulation calculation, pass a calculation function into the plotter call. Data for the calculation function will be passed along as void pointer. The function specified in the plotter call gets a function provided to submit the results to the plotter.